MAINT: codegen: Move sympydiff_utils and custom_autowrap to separate subdirectory
